\name{contrasting-package}
\alias{contrasting-package}
\alias{contrasting}
\docType{package}
\title{
\packageTitle{contrasting}
}
\description{
\packageDescription{contrasting}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{contrasting}
\packageIndices{contrasting}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{contrasting}

Maintainer: \packageMaintainer{contrasting}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
